With how these are stored, the Y value should be a short anyways.

+++ API/Entities/BlockEntity.cs
@@ -6,7 +6,7 @@ public abstract class BlockEntity
 {
     public int Id;
     public int X;
-    public int Y;
+    public short Y;
     public int Z;
     public NbtCompound? Data;
 }
